Move AppNavHost and Screen class to navigation folder
