+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/MainActivity.kt
@@ -11,7 +11,7 @@ import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.Scaffold
 import androidx.compose.ui.Modifier
-import com.abhijith.animex.ui.screens.AppNavHost
+import com.abhijith.animex.ui.navigation.AppNavHost
 import com.abhijith.animex.ui.theme.AnimeXTheme
 
 class MainActivity : ComponentActivity() {
