+++ Abhi-AnimeX/app/src/main/java/com/abhijith/animex/ui/navigation/Screen.kt
@@ -1,4 +1,4 @@
-package com.abhijith.animex.ui.screens
+package com.abhijith.animex.ui.navigation
 
 sealed class Screen(val route: String) {
     data object AnimeList : Screen("animeList")
